Crate zng_view::webrender

Webrender build used in the view-process. A GPU based renderer for the web.

It serves as an experimental render backend for Servo, but it can also be used as such in a standalone application.

External dependencies

WebRender currently depends on FreeType

Api Structure

The main entry point to WebRender is the crate::Renderer.

By calling Renderer::new(...) you get a Renderer, as well as a RenderApiSender. Your Renderer is responsible to render the previously processed frames onto the screen.

By calling yourRenderApiSender.create_api(), you’ll get a RenderApi instance, which is responsible for managing resources and documents. A worker thread is used internally to untie the workload from the application thread and therefore be able to make better use of multicore systems.

Frame

What is referred to as a frame, is the current geometry on the screen. A new Frame is created by calling set_display_list() on the RenderApi. When the geometry is processed, the application will be informed via a RenderNotifier, a callback which you pass to [Renderer::new]. More information about stacking contexts.

set_display_list() also needs to be supplied with BuiltDisplayLists. These are obtained by finalizing a DisplayListBuilder. These are used to draw your geometry. But it doesn’t only contain trivial geometry, it can also store another StackingContext, as they’re nestable.

Modules

api

The webrender_api crate contains an assortment types and functions used by WebRender consumers as well as, in many cases, WebRender itself.

euclid

A collection of strongly typed math tools for computer graphics with an inclination towards 2d graphics and layout.

glyph_rasterizer

A glyph rasterizer for webrender

intern

The interning module provides a generic data structure interning container. It is similar in concept to a traditional string interning container, but it is specialized to the WR thread model.

render_api

shader_source

Macros

c_str

enumerate_interners

Meta-macro to enumerate the various interner identifiers and types.

Structs

AddBlobImage

Creates a blob-image resource with provided parameters.

AddFontInstance

Creates a font instance resource.

AddImage

Creates an image resource with provided parameters.

AsyncScreenshotHandle

A handle to a screenshot that is being asynchronously captured and scaled.

CaptureBits

Bit flags for WR stages to store in a capture.

CapturedDocument

Information about a loaded capture of each document that is returned by RenderBackend.

ClearCache

Mask for clearing caches in debug commands.

CompositorCapabilities

CpuProfile

DebugFlags

Flags to enable/disable various builtin debugging tools.

Device

DirtyTileDebugInfo

Debug information about a tile that was dirty and was rasterized

DocumentTransaction

FormatDesc

GpuProfile

GraphicsApiInfo

ItemUid

A globally, unique identifier

MappedTileInfo

Information about the underlying data buffer of a mapped tile.

MemoryReport

Collection of heap sizes, in bytes. cbindgen:derive-eq=false cbindgen:derive-ostream=false

NativeSurfaceId

An arbitrary identifier for a native (OS compositor) surface

NativeSurfaceInfo

Information about a bound surface that the native compositor returns to WR.

NativeTileId

PictureCacheDebugInfo

Debug information about a set of picture cache slices, exposed via RenderResults

PipelineInfo

ProgramBinary

ProgramCache

ProgramSourceDigest

RecordedFrameHandle

A handle to a recorded frame that was captured.

RenderApi

The main entry point to interact with WebRender.

RenderApiSender

Allows the API to communicate with WebRender.

RenderResults

Return type from render(), which contains some repr(C) statistics as well as some non-repr(C) data.

Renderer

The renderer is responsible for submitting to the GPU the work prepared by the RenderBackend.

RendererStats

Some basic statistics about the rendered scene, used in Gecko, as well as in wrench reftests to ensure that tests are batching and/or allocating on render targets as we expect them to.

SWGLCompositeSurfaceInfo

Descriptor for a locked surface that will be directly composited by SWGL.

ShaderPrecacheFlags

Flags that control how shaders are pre-cached, if at all.

Shaders

SharedHitTester

SliceDebugInfo

Debug information about a set of picture cache tiles, exposed via RenderResults

TextureCacheConfig

Startup parameters for the texture cache.

TileDescriptor

Uniquely describes the content of this tile, in a way that can be (reasonably) efficiently hashed and compared.

TileId

A stable ID for a given tile, to help debugging. These are also used as unique identifiers for tile surfaces when using a native compositor.

TileNode

A node in the dirty rect tracking quadtree.

Transaction

A Transaction is a group of commands to apply atomically to a document.

TransactionMsg

Represents a transaction in the format sent through the channel.

UpdateBlobImage

Updates an already existing blob-image resource.

UpdateImage

Updates an already existing image resource.

WebRenderOptions

WindowVisibility

Enums

AddFont

Creates a font resource.

ApiMsg

Message sent by the RenderApi to the render backend thread.

CompareHelperResult

Optional extra information returned by is_same when logging is enabled.

CompositorConfig

Public interface specified in WebRenderOptions that configures how WR compositing will operate.

DebugCommand

Update of the state of built-in debugging facilities.

FrameMsg

Frame messages affect frame generation (applied after building the scene).

GenerateFrame

Whether to generate a frame, and if so, an id that allows tracking this transaction through the various frame stages.

GraphicsApi

InvalidationReason

Debugging information about why a tile was invalidated

PrimitiveCompareResult

The result of a primitive dependency comparison. Size is a u8 since this is a hot path in the code, and keeping the data small is a performance win.

RendererError

ResourceUpdate

Update of a persistent resource in WebRender.

SceneMsg

Frame messages affect building the scene.

ShaderKind

TileDebugInfo

Debug information about the state of a tile

TileNodeKind

Details for a node in a quadtree that tracks dirty rects for a tile.

UploadMethod

Method of uploading texel data from CPU to GPU.

VertexUsageHint

Constants

MAX_VERTEX_TEXTURE_WIDTH

ONE_TIME_USAGE_HINT

Use this hint for all vertex data re-initialization. This allows the driver to better re-use RBOs internally.

Traits

AsyncPropertySampler

Allows callers to hook into the main render_backend loop and provide additional frame ops for generate_frame transactions. These functions are all called from the render backend thread.

Compositor

Defines an interface to a native (OS level) compositor. If supplied by the client application, then picture cache slices will be composited by the OS compositor, rather than drawn via WR batches.

MappableCompositor

A Compositor variant that supports mapping tiles into CPU memory.

PartialPresentCompositor

Defines an interface to a non-native (application-level) Compositor which handles partial present. This is required if webrender must query the backbuffer’s age. TODO: Use the Compositor trait for native and non-native compositors, and integrate this functionality there.

ProfilerHooks

Defines the interface for hooking up an external profiler to WR.

ProgramCacheObserver

The interfaces that an application can implement to handle ProgramCache update

SceneBuilderHooks

Allows callers to hook in at certain points of the async scene build. These functions are all called from the scene builder thread.

Functions

create_webrender_instance

Initializes WebRender and creates a Renderer and RenderApiSender.

get_gl_target

get_unoptimized_shader_source

set_profiler_hooks

Set the profiler callbacks, or None to disable the profiler. This function must only ever be called before any WR instances have been created, or the hooks will not be set.

Type Aliases

CompositorSurfaceTransform

The transform type to apply to Compositor surfaces.

FastHashMap

SharedShaders

TileOffset