Crate zng_view::webrender::api

The webrender_api crate contains an assortment types and functions used by WebRender consumers as well as, in many cases, WebRender itself.

This separation allows Servo to parallelize compilation across webrender and other crates that depend on webrender_api. So in practice, we put things in this crate when Servo needs to use them. Firefox depends on the webrender crate directly, and so this distinction is not really relevant there.

Modules

channel

units

A collection of coordinate spaces and their corresponding Point, Size and Rect types.

Structs

AuxIter

BackdropFilterDisplayItem

Renders a filtered region of its backdrop

BlendPrimitive

BlobImageDescriptor

Metadata (but not storage) for a blob image.

BlobImageKey

An opaque identifier describing a blob image registered with WebRender. This is used as a handle to reference blob images, and can be used as an image in display items.

BlobImageParams

Input parameters for the BlobImageRasterizer.

BlobImageRequest

A key identifying blob image rasterization work requested from the blob image rasterizer.

BlurPrimitive

BorderDisplayItem

BorderRadius

BorderSide

BoxShadowDisplayItem

BuiltDisplayList

A display list.

BuiltDisplayListDescriptor

Describes the memory layout of a display list.

BuiltDisplayListIter

ClearRectangleDisplayItem

Clears all colors from the area, making it possible to cut holes in the window. (useful for things like the macos frosted-glass effect).

ClipChainId

ClipChainItem

Just an abstraction for bundling up a bunch of clips into a “super clip”.

ClipId

A reference to a clipping node defining how an item is clipped.

ColorF

Represents RGBA screen colors with floating point numbers.

ColorMatrixPrimitive

cbindgen:derive-eq=false

ColorU

Represents RGBA screen colors with one byte per channel.

CommonItemProperties

A grouping of fields a lot of display items need, just to avoid repeating these over and over in this file.

ComplexClipRegion

ComponentTransferPrimitive

CompositePrimitive

ConicGradient

ConicGradientDisplayItem

CrashAnnotatorGuard

Guard to add a crash annotation at creation, and clear it at destruction.

DebugFlags

Flags to enable/disable various builtin debugging tools.

DisplayItemCache

DisplayItemRef

DisplayListBuilder

DisplayListPayload

DisplayListWithCache

DocumentId

A key uniquely identifying a WebRender document.

DropShadowPrimitive

DynamicProperties

When using generate_frame(), a list of PropertyValue structures can optionally be supplied to provide the current value of any animated properties.

Epoch

An epoch identifies the state of a pipeline in time.

ExternalEvent

An opaque pointer-sized value.

ExternalImage

The data that an external client should provide about an external image. For instance, if providing video frames, the application could call wr.render() whenever a new video frame is ready. Note that the UV coords are supplied in texel-space!

ExternalImageData

Descriptor for external image resources. See ImageData.

ExternalImageId

An arbitrary identifier for an external image provided by the application. It must be a unique identifier for each external image.

ExternalScrollId

An external identifier that uniquely identifies a scroll frame independent of its ClipId, which may change from frame to frame. This should be unique within a pipeline. WebRender makes no attempt to ensure uniqueness. The zero value is reserved for use by the root scroll node of every pipeline, which always has an external id.

FilterData

FilterPrimitive

SVG Filter Primitive. See: https://github.com/eqrion/cbindgen/issues/9 cbindgen:derive-eq=false

FloodPrimitive

FontInstanceData

Data corresponding to an instantiation of a font, with size and other options specified.

FontInstanceFlags

FontInstanceKey

FontInstanceOptions

FontInstancePlatformOptions

FontKey

FontSize

Hashable floating-point storage for font size.

FontVariation

FramePublishId

GlyphDimensionRequest

GlyphDimensions

GlyphIndexRequest

GlyphInstance

GlyphOptions

Gradient

GradientBuilder

Construct a gradient to be used in display lists.

GradientDisplayItem

The area

GradientStop

HitTestDisplayItem

A minimal hit-testable item for the parent browser’s convenience, and is slimmer than a RectangleDisplayItem (no color). The existence of this as a distinct item also makes it easier to inspect/debug display items.

HitTestResult

Returned by RenderApi::hit_test.

HitTestResultItem

Describe an item that matched a hit-test query.

HitTesterRequest

A hit tester requested to the render backend thread but not necessarily ready yet.

IdNamespace

ID namespaces uniquely identify different users of WebRender’s API.

IdentityPrimitive

IframeDisplayItem

ImageDescriptor

Metadata (but not storage) describing an image In WebRender.

ImageDescriptorFlags

Various flags that are part of an image descriptor.

ImageDisplayItem

This describes an image that fills the specified area. It stretches or shrinks the image as necessary. While RepeatingImageDisplayItem could otherwise provide a superset of the functionality, it has been problematic inferring the desired repetition properties when snapping changes the size of the primitive.

ImageKey

An opaque identifier describing an image registered with WebRender. This is used as a handle to reference images, and is used as the hash map key for the actual image storage in the ResourceCache.

ImageMask

ImageMaskClipDisplayItem

ItemRange

ItemStats

Stats for an individual item

LineDisplayItem

MinimapData

NativeFontHandle

NinePatchBorder

NormalBorder

NotificationRequest

A request to notify a handler when the transaction reaches certain stages of the rendering pipeline.

OffsetPrimitive

OpacityPrimitive

PipelineId

From the point of view of WR, PipelineId is completely opaque and generic as long as it’s clonable, serializable, comparable, and hashable.

PremultipliedColorF

Represents pre-multiplied RGBA colors with floating point numbers.

PrimitiveFlags

PropertyBindingId

A key to identify an animated property binding.

PropertyBindingKey

A unique key that is used for connecting animated property values to bindings in the display list.

PropertyValue

The current value of an animated property. This is supplied by the calling code.

PushShadowDisplayItem

PushStackingContextDisplayItem

QualitySettings

Various settings that the caller can select based on desired tradeoffs between rendering quality and performance / power usage.

RadialGradient

RadialGradientDisplayItem

RasterizedBlobImage

Representation of a rasterized blob image. This is obtained by passing BlobImageData to the embedding via the rasterization callback.

RectClipDisplayItem

RectangleDisplayItem

A solid or an animating color to draw (may not actually be a rectangle due to complex clips)

ReferenceFrame

ReferenceFrameDescriptor

ReferenceFrameDisplayListItem

ReferenceFrameMapper

Maps from stacking context layout coordinates into reference frame relative coordinates.

RenderReasons

Flags to track why we are rendering.

RepeatingImageDisplayItem

This describes a background-image and its tiling. It repeats in a grid to fill the specified area.

RoundedRectClipDisplayItem

SampledScrollOffset

SaveState

ScrollFrameDescriptor

Shadow

SpaceAndClipInfo

Per-primitive information about the nodes in the clip tree and the spatial tree that the primitive belongs to.

SpatialId

A reference to a spatial node defining item positioning.

SpatialNodeInfo

A small portion of a normal spatial node that we store during DL construction to enable snapping and reference frame <-> stacking context coord mapping. In future we’ll aim to remove this and have the full spatial tree available during DL build.

SpatialTreeItemKey

Defines a caller provided key that is unique for a given spatial node, and is stable across display lists. WR uses this to determine which spatial nodes are added / removed for a new display list. The content itself is arbitrary and opaque to WR, the only thing that matters is that it’s unique and stable between display lists.

StackingContext

StackingContextFlags

StickyFrameDescriptor

StickyOffsetBounds

The minimum and maximum allowable offset for a sticky frame in a single dimension.

SyntheticItalics

TempFilterData

TextDisplayItem

YuvImageDisplayItem

Enums

AlphaType

BlobImageError

Error code for when blob rasterization failed.

BoolParameter

Boolean configuration option.

BorderDetails

BorderRadiusKind

BorderStyle

BoxShadowClipMode

Checkpoint

A stage of the rendering pipeline.

ClipMode

ColorDepth

Specifies the color depth of an image. Currently only used for YUV images.

ColorRange

ColorSpace

ComponentTransferFuncType

CompositeOperator

Available composite operoations for the composite filter primitive

CrashAnnotation

Crash annotations included in crash reports.

DirtyRect

The possible states of a Dirty rect.

DisplayItem

DisplayListSection

DisplayListSection determines the target buffer for the display items.

ExtendMode

ExternalImageSource

The source for an external image.

ExternalImageType

Storage format identifier for externally-managed images.

FillRule

FilterOp

CSS filter.

FilterPrimitiveInput

An input to a SVG filter primitive.

FilterPrimitiveKind

See: https://github.com/eqrion/cbindgen/issues/9 cbindgen:derive-eq=false

FontHinting

FontLCDFilter

FontRenderMode

FontTemplate

Container for the raw data describing a font. This might be a stream of bytes corresponding to a downloaded font, or a handle to a native font from the operating system.

GeckoDisplayListType

HasScrollLinkedEffect

A flag in each scrollable frame to represent whether the owner of the frame document has any scroll-linked effect. See https://firefox-source-docs.mozilla.org/performance/scroll-linked_effects.html for a definition of scroll-linked effect.

ImageBufferKind

Specifies the type of texture target in driver terms.

ImageData

Represents the backing store of an arbitrary series of pixels for display by WebRender. This storage can take several forms.

ImageFormat

Specifies the format of a series of pixels, in driver terms.

ImageRendering

IntParameter

Integer configuration option.

LineOrientation

LineStyle

MixBlendMode

NinePatchBorderSource

Parameter

A configuration option that can be changed at runtime.

PrimitiveKeyKind

Information specific to a primitive type that uniquely identifies a primitive template by key.

PropertyBinding

A binding property can either be a specific value (the normal, non-animated case) or point to a binding location to fetch the current value from. Note that Binding has also a non-animated value, the value is used for the case where the animation is still in-delay phase (i.e. the animation doesn’t produce any animation values).

RasterSpace

Configure whether the contents of a stacking context should be rasterized in local space or screen space. Local space rasterized pictures are typically used when we want to cache the output, and performance is important. Note that this is a performance hint only, which WR may choose to ignore.

ReferenceFrameKind

ReferenceTransformBinding

RepeatMode

Rotation

ScrollLocation

SpatialTreeItem

TransformStyle

YuvColorSpace

YuvData

YuvFormat

YuvRangedColorSpace

Constants

DEFAULT_TILE_SIZE

The default tile size for blob images and regular images larger than the maximum texture size.

MAX_RENDER_TASK_SIZE

Defined here for cbindgen

MAX_TEXT_RUN_LENGTH

POLYGON_CLIP_VERTEX_MAX

Traits

ApiHitTester

An object that can perform hit-testing without doing synchronous queries to the RenderBackendThread.

AsyncBlobImageRasterizer

A group of rasterization requests to execute synchronously on the scene builder thread.

BlobImageHandler

A handler on the render backend that can create rasterizer objects which will be sent to the scene builder thread to execute the rasterization.

BlobImageResources

The resources exposed by the resource cache available for use by the blob rasterizer.

CrashAnnotator

Handler to expose support for annotating crash reports.

ExternalImageHandler

The interfaces that an application can implement to support providing external image buffers. When the application passes an external image to WR, it should keep that external image life time. People could check the epoch id in RenderNotifier at the client side to make sure that the external image is not used by WR. Then, do the clean up for that external image.

NotificationHandler

A handler to notify when a transaction reaches certain stages of the rendering pipeline.

RenderNotifier

A handler to integrate WebRender with the thread that contains the Renderer.

Type Aliases

APZScrollGeneration

BlobImageData

Backing store for blob image command streams.

BlobImageResult

Result type for blob raserization.

GlyphIndex

ItemKey

An identifier used to refer to previously sent display items. Currently it refers to individual display items, but this may change later.

ItemTag

A tag that can be used to identify items during hit testing. If the tag is missing then the item doesn’t take part in hit testing at all. This is composed of two numbers. In Servo, the first is an identifier while the second is used to select the cursor that should be used during mouse movement. In Gecko, the first is a scrollframe identifier, while the second is used to store various flags that APZ needs to properly process input events.

PipelineSourceId

This type carries no valuable semantics for WR. However, it reflects the fact that clients (Servo) may generate pipelines by different semi-independent sources. These pipelines still belong to the same IdNamespace and the same DocumentId. Having this extra Id field enables them to generate PipelineId without collision.

TileSize

Width and height in device pixels of image tiles.

VoidPtrToSizeFn

A C function that takes a pointer to a heap allocation and returns its size.