Type Alias zng_view::webrender::api::units::LayoutSize

pub type LayoutSize = Size2D<f32, LayoutPixel>;

Aliased Type

struct LayoutSize {
    pub width: f32,
    pub height: f32,
    /* private fields */
}

Fields
width: f32

The extent of the element in the U units along the x axis (usually horizontal).

height: f32

The extent of the element in the U units along the y axis (usually vertical).