Struct zng_view::webrender::euclid::Box2D

#[repr(C)]
pub struct Box2D<T, U> {
    pub min: Point2D<T, U>,
    pub max: Point2D<T, U>,
}

A 2d axis aligned rectangle represented by its minimum and maximum coordinates.

Representation

This struct is similar to Rect, but stores rectangle as two endpoints instead of origin point and size. Such representation has several advantages over Rect representation:

Several operations are more efficient with Box2D, including intersection, union, and point-in-rect.

The representation is less susceptible to overflow. With Rect, computation of second point can overflow for a large range of values of origin and size. However, with Box2D, computation of size cannot overflow if the coordinates are signed and the resulting size is unsigned.

A known disadvantage of Box2D is that translating the rectangle requires translating both points, whereas translating Rect only requires translating one point.

Empty box

A box is considered empty (see is_empty) if any of the following is true:

it’s area is empty,

it’s area is negative (min.x > max.x or min.y > max.y),

it contains NaNs.

Fields
min: Point2D<T, U>max: Point2D<T, U>
Implementations

impl<T, U> Box2D<T, U>

pub const fn new(min: Point2D<T, U>, max: Point2D<T, U>) -> Box2D<T, U>

Constructor.

pub fn from_origin_and_size( origin: Point2D<T, U>, size: Size2D<T, U> ) -> Box2D<T, U>
where T: Copy + Add<Output = T>,

Constructor.

pub fn from_size(size: Size2D<T, U>) -> Box2D<T, U>
where T: Zero,

Creates a Box2D of the given size, at offset zero.

impl<T, U> Box2D<T, U>
where T: PartialOrd,

pub fn is_negative(&self) -> bool

Returns true if the box has a negative area.

The common interpretation for a negative box is to consider it empty. It can be obtained by calculating the intersection of two boxes that do not intersect.

pub fn is_empty(&self) -> bool

Returns true if the size is zero, negative or NaN.

pub fn intersects(&self, other: &Box2D<T, U>) -> bool

Returns true if the two boxes intersect.

pub fn contains(&self, p: Point2D<T, U>) -> bool

Returns true if this box contains the point. Points are considered in the box if they are on the front, left or top faces, but outside if they are on the back, right or bottom faces.

pub fn contains_box(&self, other: &Box2D<T, U>) -> bool

Returns true if this box contains the interior of the other box. Always returns true if other is empty, and always returns false if other is nonempty but this box is empty.

impl<T, U> Box2D<T, U>
where T: Copy + PartialOrd,

pub fn to_non_empty(&self) -> Option<Box2D<T, U>>

pub fn intersection(&self, other: &Box2D<T, U>) -> Option<Box2D<T, U>>

Computes the intersection of two boxes, returning None if the boxes do not intersect.

pub fn intersection_unchecked(&self, other: &Box2D<T, U>) -> Box2D<T, U>

Computes the intersection of two boxes without check whether they do intersect.

The result is a negative box if the boxes do not intersect. This can be useful for computing the intersection of more than two boxes, as it is possible to chain multiple intersection_unchecked calls and check for empty/negative result at the end.

pub fn union(&self, other: &Box2D<T, U>) -> Box2D<T, U>

Computes the union of two boxes.

If either of the boxes is empty, the other one is returned.

impl<T, U> Box2D<T, U>
where T: Copy + Add<Output = T>,

pub fn translate(&self, by: Vector2D<T, U>) -> Box2D<T, U>

Returns the same box, translated by a vector.

impl<T, U> Box2D<T, U>
where T: Copy + Sub<Output = T>,

pub fn size(&self) -> Size2D<T, U>

pub fn set_size(&mut self, size: Size2D<T, U>)

Change the size of the box by adjusting the max endpoint without modifying the min endpoint.

pub fn width(&self) -> T

pub fn height(&self) -> T

pub fn to_rect(&self) -> Rect<T, U>

impl<T, U> Box2D<T, U>
where T: Copy + Add<Output = T> + Sub<Output = T>,

pub fn inflate(&self, width: T, height: T) -> Box2D<T, U>

Inflates the box by the specified sizes on each side respectively.

pub fn inner_box(&self, offsets: SideOffsets2D<T, U>) -> Box2D<T, U>

Calculate the size and position of an inner box.

Subtracts the side offsets from all sides. The horizontal, vertical and applicate offsets must not be larger than the original side length.

pub fn outer_box(&self, offsets: SideOffsets2D<T, U>) -> Box2D<T, U>

Calculate the b and position of an outer box.

Add the offsets to all sides. The expanded box is returned.

impl<T, U> Box2D<T, U>
where T: Copy + Zero + PartialOrd,

pub fn from_points<I>(points: I) -> Box2D<T, U>
where I: IntoIterator, <I as IntoIterator>::Item: Borrow<Point2D<T, U>>,

Returns the smallest box containing all of the provided points.

impl<T, U> Box2D<T, U>
where T: One + Add<Output = T> + Sub<Output = T> + Mul<Output = T> + Copy,

pub fn lerp(&self, other: Box2D<T, U>, t: T) -> Box2D<T, U>

Linearly interpolate between this box and another box.

impl<T, U> Box2D<T, U>
where T: Copy + One + Add<Output = T> + Div<Output = T>,

pub fn center(&self) -> Point2D<T, U>

impl<T, U> Box2D<T, U>
where T: Copy + Mul<Output = T> + Sub<Output = T>,

pub fn area(&self) -> T

impl<T, U> Box2D<T, U>
where T: Zero,

pub fn zero() -> Box2D<T, U>

Constructor, setting all sides to zero.

impl<T, U> Box2D<T, U>
where T: Copy,

pub fn x_range(&self) -> Range<T>

pub fn y_range(&self) -> Range<T>

pub fn to_untyped(&self) -> Box2D<T, UnknownUnit>

Drop the units, preserving only the numeric value.

pub fn from_untyped(c: &Box2D<T, UnknownUnit>) -> Box2D<T, U>

Tag a unitless value with units.

pub fn cast_unit<V>(&self) -> Box2D<T, V>

Cast the unit

pub fn scale<S>(&self, x: S, y: S) -> Box2D<T, U>
where S: Copy, T: Mul<S, Output = T>,

impl<T, U> Box2D<T, U>
where T: NumCast + Copy,

pub fn cast<NewT>(&self) -> Box2D<NewT, U>
where NewT: NumCast,

Cast from one numeric representation to another, preserving the units.

When casting from floating point to integer coordinates, the decimals are truncated as one would expect from a simple cast, but this behavior does not always make sense geometrically. Consider using round(), round_in or round_out() before casting.

pub fn try_cast<NewT>(&self) -> Option<Box2D<NewT, U>>
where NewT: NumCast,

Fallible cast from one numeric representation to another, preserving the units.

When casting from floating point to integer coordinates, the decimals are truncated as one would expect from a simple cast, but this behavior does not always make sense geometrically. Consider using round(), round_in or round_out() before casting.

pub fn to_f32(&self) -> Box2D<f32, U>

Cast into an f32 box.

pub fn to_f64(&self) -> Box2D<f64, U>

Cast into an f64 box.

pub fn to_usize(&self) -> Box2D<usize, U>

Cast into an usize box, truncating decimals if any.

When casting from floating point boxes, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

pub fn to_u32(&self) -> Box2D<u32, U>

Cast into an u32 box, truncating decimals if any.

When casting from floating point boxes, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

pub fn to_i32(&self) -> Box2D<i32, U>

Cast into an i32 box, truncating decimals if any.

When casting from floating point boxes, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

pub fn to_i64(&self) -> Box2D<i64, U>

Cast into an i64 box, truncating decimals if any.

When casting from floating point boxes, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

impl<T, U> Box2D<T, U>
where T: Float,

pub fn is_finite(self) -> bool

Returns true if all members are finite.

impl<T, U> Box2D<T, U>
where T: Round,

pub fn round(&self) -> Box2D<T, U>

Return a box with edges rounded to integer coordinates, such that the returned box has the same set of pixel centers as the original one. Values equal to 0.5 round up. Suitable for most places where integral device coordinates are needed, but note that any translation should be applied first to avoid pixel rounding errors. Note that this is not rounding to nearest integer if the values are negative. They are always rounding as floor(n + 0.5).

impl<T, U> Box2D<T, U>
where T: Floor + Ceil,

pub fn round_in(&self) -> Box2D<T, U>

Return a box with faces/edges rounded to integer coordinates, such that the original box contains the resulting box.

pub fn round_out(&self) -> Box2D<T, U>

Return a box with faces/edges rounded to integer coordinates, such that the original box is contained in the resulting box.