Module zng_view::webrender::render_api

Structs

AddBlobImage

Creates a blob-image resource with provided parameters.

AddFontInstance

Creates a font instance resource.

AddImage

Creates an image resource with provided parameters.

CaptureBits

Bit flags for WR stages to store in a capture.

CapturedDocument

Information about a loaded capture of each document that is returned by RenderBackend.

ClearCache

Mask for clearing caches in debug commands.

DocumentTransaction

MemoryReport

Collection of heap sizes, in bytes. cbindgen:derive-eq=false cbindgen:derive-ostream=false

RenderApi

The main entry point to interact with WebRender.

RenderApiSender

Allows the API to communicate with WebRender.

Transaction

A Transaction is a group of commands to apply atomically to a document.

TransactionMsg

Represents a transaction in the format sent through the channel.

UpdateBlobImage

Updates an already existing blob-image resource.

UpdateImage

Updates an already existing image resource.

Enums

AddFont

Creates a font resource.

ApiMsg

Message sent by the RenderApi to the render backend thread.

DebugCommand

Update of the state of built-in debugging facilities.

FrameMsg

Frame messages affect frame generation (applied after building the scene).

GenerateFrame

Whether to generate a frame, and if so, an id that allows tracking this transaction through the various frame stages.

ResourceUpdate

Update of a persistent resource in WebRender.

SceneMsg

Frame messages affect building the scene.