zng_view

Module extensions

Extensions API

Extensions that run in the view-process, with internal access to things like the raw handle of windows or direct access to renderers. These extensions are build on top of the view API extensions as a way to customize the view-process without needing to fork it or re-implement the entire view API from scratch.

Structs

BlobAddArgs

Arguments for BlobExtension::add.

BlobPrepareArgs

Arguments for BlobExtension::prepare_resources.

BlobRasterizerArgs

Arguments for AsyncBlobRasterizer::rasterize.

BlobUpdateArgs

Arguments for BlobExtension::update.

ExtensionEventSender

Sender of Event::ExtensionEvent events.

ExternalImages

Tracks extension external images for a renderer.

FrameReadyArgs

Arguments for RendererExtension::frame_ready.

RedrawArgs

Arguments for RendererExtension::redraw.

RenderArgs

Arguments for RendererExtension::render_start and RendererExtension::render_end.

RenderItemArgs

Arguments for RendererExtension::render_push and RendererExtension::render_pop.

RenderUpdateArgs

Arguments for RendererExtension::render_update.

RendererCommandArgs

Arguments for RendererExtension::command.

RendererConfigArgs

Arguments for RendererExtension::configure

RendererDeinitedArgs

Arguments for RendererExtension::renderer_deinited.

RendererInitedArgs

Arguments for RendererExtension::renderer_inited.

ViewExtensionInitArgs

Arguments for ViewExtension::init.

ViewExtensions

View extensions register.

WindowCommandArgs

Arguments for WindowExtension::command.

WindowConfigArgs

Arguments for WindowExtension::configure

WindowDeinitedArgs

Arguments for WindowExtension::window_deinited.

WindowEventArgs

Arguments for WindowExtension::event.

WindowInitedArgs

Arguments for WindowExtension::window_inited.

Traits

AsyncBlobRasterizer

Snapshot of a BlobExtension that can render/copy pixels.

BlobExtension

Represents a Webrender blob handler that can coexist with other blob handlers on the same renderer.

OpenGlContext

Represents a managed OpenGL context connected to a window or headless surface.

PxToWr

Conversion from Px to webrender units.

RendererExtension

Represents a view extension associated with a renderer instance.

ViewExtension

The extension API.

WindowExtension

Represents a view extension associated with a headed or headless window instance.

WrToPx

Conversion from webrender to Px units.