Trait zng_view::extensions::RendererExtension

pub trait RendererExtension: Any {

    // Required method
    fn is_config_only(&self) -> bool;

    // Provided methods
    fn configure(&mut self, args: &mut RendererConfigArgs<'_>) { ... }
    fn renderer_inited(&mut self, args: &mut RendererInitedArgs<'_>) { ... }
    fn renderer_deinited(&mut self, args: &mut RendererDeinitedArgs<'_>) { ... }
    fn command(
        &mut self,
        args: &mut RendererCommandArgs<'_>
    ) -> ApiExtensionPayload { ... }
    fn render_start(&mut self, args: &mut RenderArgs<'_>) { ... }
    fn render_push(&mut self, args: &mut RenderItemArgs<'_>) { ... }
    fn render_pop(&mut self, args: &mut RenderItemArgs<'_>) { ... }
    fn render_end(&mut self, args: &mut RenderArgs<'_>) { ... }
    fn render_update(&mut self, args: &mut RenderUpdateArgs<'_>) { ... }
    fn frame_ready(&mut self, args: &mut FrameReadyArgs) { ... }
    fn redraw(&mut self, args: &mut RedrawArgs<'_>) { ... }
    fn low_memory(&mut self) { ... }

}

Represents a view extension associated with a renderer instance.

Required Methods

fn is_config_only(&self) -> bool

If this extension can be dropped after render creation.

Provided Methods

fn configure(&mut self, args: &mut RendererConfigArgs<'_>)

Edit options for the new renderer.

fn renderer_inited(&mut self, args: &mut RendererInitedArgs<'_>)

Called just after the renderer is created.

fn renderer_deinited(&mut self, args: &mut RendererDeinitedArgs<'_>)

Called just before the renderer is destroyed.

fn command(&mut self, args: &mut RendererCommandArgs<'_>) -> ApiExtensionPayload

Called when a command request is made for the extension and renderer (window ID).

The extension_id is the current index of the extension, it can be used in error messages.

fn render_start(&mut self, args: &mut RenderArgs<'_>)

Called when a new display list begins building.

fn render_push(&mut self, args: &mut RenderItemArgs<'_>)

Called when a display item push for the extension is found.

fn render_pop(&mut self, args: &mut RenderItemArgs<'_>)

Called when a display item pop for the extension is found.

fn render_end(&mut self, args: &mut RenderArgs<'_>)

Called when a new display list finishes building.

The list will be send to the renderer for asynchronous rendering.

fn render_update(&mut self, args: &mut RenderUpdateArgs<'_>)

Called when a render-update for the extension is found.

fn frame_ready(&mut self, args: &mut FrameReadyArgs)

Called when Webrender finishes rendering a frame and it is ready for redraw.

fn redraw(&mut self, args: &mut RedrawArgs<'_>)

Called every time the window or surface redraws, after Webrender has redraw.

fn low_memory(&mut self)

System warning low memory, release unused memory, caches.