Trait zng_view::extensions::WindowExtension

pub trait WindowExtension: Any {
    // Required method
    fn is_init_only(&self) -> bool;

    // Provided methods
    fn configure(&mut self, args: &mut WindowConfigArgs<'_>) { ... }
    fn window_inited(&mut self, args: &mut WindowInitedArgs<'_>) { ... }
    fn command(
        &mut self,
        args: &mut WindowCommandArgs<'_>
    ) -> ApiExtensionPayload { ... }
    fn event(&mut self, args: &mut WindowEventArgs<'_>) { ... }
    fn low_memory(&mut self) { ... }
    fn window_deinited(&mut self, args: &mut WindowDeinitedArgs<'_>) { ... }
}

Represents a view extension associated with a headed window instance.

Required Methods

fn is_init_only(&self) -> bool

If this extension can be dropped after window creation.

Provided Methods

fn configure(&mut self, args: &mut WindowConfigArgs<'_>)

Edit attributes for the new window.

fn window_inited(&mut self, args: &mut WindowInitedArgs<'_>)

Called just after the window is created.

fn command(&mut self, args: &mut WindowCommandArgs<'_>) -> ApiExtensionPayload

Called when a command request is made for the extension and window (window ID).

fn event(&mut self, args: &mut WindowEventArgs<'_>)

Called when the window receives an event.

fn low_memory(&mut self)

System warning low memory, release unused memory, caches.

fn window_deinited(&mut self, args: &mut WindowDeinitedArgs<'_>)

Called just after the window closes.